+++ src/main/java/com/hnly/provincial/service/wateruserecords/IWaterUseRecordsService.java
@@ -80,5 +80,5 @@ public interface IWaterUseRecordsService extends IService<WaterUseRecords> {
      * @param code 区域
      * @return 该区域该年每月的累计用水量
      */
-    List<BigDecimal> getMonthSumWaterByYear(Integer year, String code);
+    List<BigDecimal> getMonthSumWaterByYear(Long year, String code);
 }
